Update Image Magnifier Glass.txt
The first task is to add the image for the logo. Steps to include image and create logo: 
 

Download image by clicking here.
Copy and Paste the image to the directory: root/images. Where root is the top directory of our project. In our case it is named as “sample project”.
Include the image in the code using img tag.
The second task is to create an unordered-list in HTML inside the nav section of the header menu: 
 

Add an unordered list in the navigation menu section with 5 list-items named “Home”, “About Us”, “Our Products”, “Careers”, and “Contact Us”.
The code of the Header section after adding the above two things will look like as shown below: